requirements.txt
requests
python-dotenv
pdfplumber
python-docx
textract==1.6.3

--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

typing-extensions==4.12.2
numpy==2.2.1

torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118

kornia>=0.6.0
Pillow>=8.0.0
python-Levenshtein>=0.12.0
jiwer>=2.3.0
tensorboard>=2.4.0
matplotlib>=3.3.0

tqdm

# для распознавания речи
git+https://github.com/openai/whisper.git
soundfile

soundfile
omegaconf

aiogram==3.13.1
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pdfminer.six==20231228
chardet==5.2.0
uvloop==0.19.0; sys_platform != 'win32'
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
aiosqlite==0.20.0   # уже есть, но на всякий случай
python-multipart==0.0.9